--- a/src/compiler/tsconfig.json
+++ b/src/compiler/tsconfig.json
@@ -12,6 +12,7 @@
     },
     "files": [
         "core.ts",
+        "dataStructures.ts",
         "performance.ts",
         "sys.ts",
         "types.ts",