… src/compiler/checker.ts
… src/compiler/core.ts
… src/compiler/dataStructures.ts
… src/compiler/emitter.ts
… src/compiler/program.ts
… src/compiler/sys.ts
… src/compiler/transformer.ts
… src/compiler/transformers/es6.ts
… src/compiler/transformers/generators.ts
… src/compiler/transformers/jsx.ts
… src/compiler/transformers/module/module.ts
… src/compiler/transformers/module/system.ts
… src/compiler/transformers/ts.ts
… src/compiler/tsc.ts
… src/compiler/types.ts
… src/compiler/utilities.ts
… src/compiler/visitor.ts
… src/harness/fourslash.ts
… src/harness/harness.ts
… src/harness/harnessLanguageService.ts
… src/harness/projectsRunner.ts
… src/harness/unittests/cachingInServerLSHost.ts
… src/harness/unittests/moduleResolution.ts
… src/harness/unittests/reuseProgramStructure.ts
… src/harness/unittests/session.ts
… src/harness/unittests/tsserverProjectSystem.ts
… src/server/client.ts
… src/server/editorServices.ts
… src/server/session.ts
… src/services/classifier.ts