@andy-ms 2016-10-04
Closing in favor of #11354, which uses maps directly (using infix `map.get(key)` syntax) rather than through an abstract data type.